brocode/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
boto3==1.26.65
botocore==1.29.139
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.7
django-cors-headers==3.8.0
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djongo==1.3.6
dnspython==2.3.0
docopt==0.6.2
Faker==15.1.3
idna==3.4
jmespath==1.0.1
kafka-python==2.0.2
newrelic==8.4.0
num2words==0.5.12
pdfkit==1.0.0
PyJWT==1.7.1
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2022.4
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.28.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.12
